base-item: Split the loading and editing graphs
Adding and removing operations to the graph in one thread and then
processing it in another is causing intermittent crashes due to what
looks like threading and/or memory issues. However, we can not move
the entire processing to the main thread either. Loading big, high-res
image files in the main thread will badly freeze the UI.

Therefore, we use a small graph (gegl:load -> gegl:buffer-sink) that
we process in a separate thread to load the image into a GeglBuffer.
Then we use the GeglBuffer to drive a separate graph with the editing
operations (gegl:buffer-source -> PhotosPipeline). This graph is
processed in an idle handler in the main thread. This gives us both
asynchronous loading of data, and the ability to reliably tinker with
the editing graph.

Using a GeglBuffer in this way to hold the two graphs together seems
like a good solution to our problem. GeglBuffers have received enough
beating to ensure that they can be used from different threads.

In case you are wondering, the GIMP uses separate plug-in processes to
load the images, which are then fed to the "core" using pipes. That is
why the issue of asynchronous loading doesn't arise there.